+++ src/region/builder/mod.rs
@@ -15,7 +15,7 @@ pub fn builder(region: &mut Region, planet: &Planet, crash_site: Point) {
     let mut pooled_water = vec![0u8; REGION_WIDTH as usize * REGION_HEIGHT as usize];
     println!("{}", planet.water_height);
     let mut rng = RandomNumberGenerator::seeded(
-        planet.perlin_seed + planet_idx(crash_site.x, crash_site.y) as u64,
+        planet.perlin_seed + planet_idx(crash_site.x as usize, crash_site.y as usize) as u64,
     );
 
     set_worldgen_status("Establishing ground altitude");
@@ -27,7 +27,7 @@ pub fn builder(region: &mut Region, planet: &Planet, crash_site: Point) {
         &mut hm,
         crash_site,
         planet.perlin_seed,
-        planet.landblocks[planet_idx(crash_site.x, crash_site.y)].variance,
+        planet.landblocks[planet_idx(crash_site.x as usize, crash_site.y as usize)].variance,
     );
     crate::planet::WORLDGEN_RENDER
         .lock()
@@ -36,7 +36,7 @@ pub fn builder(region: &mut Region, planet: &Planet, crash_site: Point) {
     set_worldgen_status("Locating Sub-Biomes");
     heightmap::create_subregions(
         &mut rng,
-        planet.landblocks[planet_idx(crash_site.x, crash_site.y)].variance,
+        planet.landblocks[planet_idx(crash_site.x as usize, crash_site.y as usize)].variance,
         &mut hm,
         &mut pooled_water,
         &biome,
